consider cost of reproduction
In this version,  females first reproduce several offspring proportional to their fitness and some stochasticity (assume the number of offspring a female with fitness x actually produce follows a poisson distribution with lambda = x). And offspring are sampled from the pool.
I kept the “fried egg” environment updating rules, soft selection on males and hard selection on females. Fitness is determined simply by exp(- 3d), in which d is the distance between the individual’s trait value and the habitat trait centre.